fix: split two components and removed some logic from render method
